Adds builds for armhf and arm64 architectures.
This also refactors the workflow a bit to only run the reprepro job
once, for all DEBs.

Since cross-builds are scary, this also adds a QEMU test job for
them, to ensure they can be installed and the binary can execute.

The two build jobs are kept separate, even if they have lots of
overlap, so that testing can start without waiting for native builds.

Thanks to

* @bensteinberg for providing his build script.
* https://github.com/kornelski/cargo-deb#cross-compilation
* https://github.com/marketplace/actions/arm-runner

Closes #2.